+++ .changeset/bright-mimes-leave.md
@@ -1,6 +1,7 @@
 ---
 "@effect/platform-node": patch
+"effect": patch
 ---
 
-Remove the `mime` runtime dependency. The `Mime` module now provides top-level lookup functions backed by a vendored
-standard MIME registry.
+Remove the `mime` runtime dependency. The new `effect/unstable/http/Mime` module provides top-level lookup functions
+backed by a vendored standard MIME registry.
