+++ packages/platform/node/src/NodeHttpPlatform.ts
@@ -18,9 +18,9 @@ import * as Headers from "effect/unstable/http/Headers"
 import * as HttpBody from "effect/unstable/http/HttpBody"
 import * as Platform from "effect/unstable/http/HttpPlatform"
 import * as ServerResponse from "effect/unstable/http/HttpServerResponse"
+import * as Mime from "effect/unstable/http/Mime"
 import * as Fs from "node:fs"
 import { Readable } from "node:stream"
-import * as Mime from "./Mime.ts"
 import * as NodeFileSystem from "./NodeFileSystem.ts"
 import * as NodeStream from "./NodeStream.ts"
